Currently most of the test files have a separate dwarf and a separate
dsym test with almost identical content (only the build step is
different). With adding dwo symbol file handling to the test suit it
would increase this to a 3-way duplication. The purpose of this change
is to eliminate this redundancy with generating 2 test case (one dwarf
and one dsym) for each test function specified (dwo handling will be
added at a later commit).

Main design goals:
* There should be no boilerplate code in each test file to support the
  multiple debug info in most of the tests (custom scenarios are
  acceptable in special cases) so adding a new test case is easier and
  we can't miss one of the debug info type.
* In case of a test failure, the debug symbols used during the test run
  have to be cleanly visible from the output of dotest.py to make
  debugging easier both from build bot logs and from local test runs
* Each test case should have a unique, fully qualified name so we can
  run exactly 1 test with "-f <test-case>.<test-function>" syntax
* Test output should be grouped based on test files the same way as it
  happens now (displaying dwarf/dsym results separately isn't
  preferable)

Proposed solution (main logic in lldbtest.py, rest of them are test
cases fixed up for the new style):
* Have only 1 test fuction in the test files what will run for all
  debug info separately and this test function should call just
  "self.build(...)" to build an inferior with the right debug info
* When a class is created by python (the class object, not the class
  instance), we will generate a new test method for each debug info
  format in the test class with the name "<test-function>_<debug-info>"
  and remove the original test method. This way unittest2 see multiple
  test methods (1 for each debug info, pretty much as of now) and will
  handle the test selection and the failure reporting correctly (the
  debug info will be visible from the end of the test name)
* Add new annotation @no_debug_info_test to disable the generation of
  multiple tests for each debug info format when the test don't have an
  inferior

Differential revision: http://reviews.llvm.org/D13028

llvm-svn: 248883

+++ lldb/test/functionalities/exec/TestExec.py
@@ -21,35 +21,19 @@ class ExecTestCase(TestBase):
 
     mydir = TestBase.compute_mydir(__file__)
 
-        
     @skipUnlessDarwin
-    @dsym_test
-    def test_with_dsym (self):
+    def test(self):
         if self.getArchitecture() == 'x86_64':
             source = os.path.join (os.getcwd(), "main.cpp")
             o_file = os.path.join (os.getcwd(), "main.o")
             execute_command ("'%s' -g -O0 -arch i386 -arch x86_64 '%s' -c -o '%s'" % (os.environ["CC"], source, o_file))
             execute_command ("'%s' -g -O0 -arch i386 -arch x86_64 '%s'" % (os.environ["CC"], o_file))
+            if self.debug_info != "dsym":
+                dsym_path = os.path.join (os.getcwd(), "a.out.dSYM")
+                execute_command ("rm -rf '%s'" % (dsym_path))
         else:
-            self.buildDsym()
-        self.do_test ()
+            self.build()
 
-
-    @skipUnlessDarwin
-    @dwarf_test
-    def test_with_dwarf (self):
-        if self.getArchitecture() == 'x86_64':
-            source = os.path.join (os.getcwd(), "main.cpp")
-            o_file = os.path.join (os.getcwd(), "main.o")
-            dsym_path = os.path.join (os.getcwd(), "a.out.dSYM")
-            execute_command ("'%s' -g -O0 -arch i386 -arch x86_64 '%s' -c -o '%s'" % (os.environ["CC"], source, o_file))
-            execute_command ("'%s' -g -O0 -arch i386 -arch x86_64 '%s'" % (os.environ["CC"], o_file))
-            execute_command ("rm -rf '%s'" % (dsym_path))
-        else:
-            self.buildDwarf()
-        self.do_test ()
-
-    def do_test (self):
         exe = os.path.join (os.getcwd(), "a.out")
         
         # Create the target
